Remove embedded_url field from Connection
Exposing the URL in the data structure doesn't make much sense, as the
URL doesn't open to a web page and instead will throw 400. The endpoint
is only serving the SDK flow.

Instead of exposing the endpoint, we stop relying on it in the
`Connection` data structure and just hardcode the endpoint in the SDK.