Use a separate view to render service name text
We are trying to support longer service name by shrinking only the
service name part of the text on the Connect Button, and can render up
to two lines of text on that part.

There should be a better way to do this, but currently I'm doing it with
two TextViews, and try to align them nicely.