Refactor text switching animation
Currently in the SDK UI, we have 3 places that need text switching transitiona animation:
1. main text (Connect <service name>
2. progress text
3. helper text

We are using ObjectAnimators now to manually animate the text switching, which is fine in terms of the animation itself. The problem comes in when the TextView needs to resize the text to fit the new text, which we will have to take into account when running the animation.

Switching to using TextSwitcher simplifies the animation quite a bit, as well as fixing the text resizing issue.