Use a custom view to wrap ViewDragHelper
Because the `iconImg` view is not a direct child of `IftttConnectButton`, processing touch events directly in the button view causes the view bound to not match, and hence some touch events are ignored.

I don't know if there is better ways to use ViewDragHelper without wrapping it with a custom view, but having a custom view to host the toggle ImageView fixes the issue.