Fix TextView related issues (#72)
* Cancel pending switch text runnable

We need to make sure if the Connect Button was re-initialized by passing in a new Connection object, we need to cancel the pending runnable.

* Set padding directly to child TextView in TextSwitcher

We have seen some inconsistent font sizes after the animation. I have not yet reproduced it consistently, but my suspesion was that we are setting the padding in the TextSwitcher, instead of the individual TextView.