SDK UI design tweaks (#85)
* Adjust various sizes

* Connect Button height
* Service icon size
* Service name maximum text size
* Helper text size
* About page title text size

* Apply different font for service names in about page

Because of the font needs to be set in the text level, we have to bring back the `AvenirTypefaceSpan`.

* Use ACTION_GO for email field

This is to make it easier for users to start the flow. We now start the flow when the user clicks the IME action button.

* Don't clip padding or children for Connect Button

This is necessary to for us to be able to increase the elevation of the knob, so that it shows a bigger shadow.

* Reduce the alpha change for email start button hit state

* Delay fade in animation

This is to make the entire cross fade animation a two-step animation, instead of cross fading both TextViews at the same time. The animation is used in the TextSwitcher.

* About page design tweaks

* Make the X button's press state visibile. The problem right now is that the theme is light, which makes the Toolbar icons' press state dark. With the black background, the press state is just not gonna be visible.

* Enable/Disable auto resizing based on space

We now always test the text space with the maximum font size. If we have enough space, disable to auto resizing and always use it. Otherwise, enable auto resizing and let it handle the sizing.

* Re-bind Connection data after the animation has completed

This is to make sure the elevation is not set while the animation is ongoing.

* Revert some changes