Consolidate revert-able error messages
We are using this kind of revert-able messages in both the helper text error messages, as well as the "slide to disconnect" message.

Instead of doing this in a one-off way (keeping a reference of the Runnable and postDelay), we can have a Handler wrapper that is dedicated to scheduling the revert action, and have a way to revert all of the pending actions at once, or just remove all of the pending actions.