Re-apply horizontal offset to views in onLayout
This might be a little bit of hack, the `ViewCompat.offsetLeftAndRight` doesn't actually change the `left` value of the view, causing it to be placed in the original place whenever there is a new layout pass.

The fix in `ButtonParentView` is to have a method to track the horizontal offset for the view, and then re-apply it to the view during `onLayout`. Theoretically we could use translationX everywhere and avoid having to do this, but the ViewDragHelper uses offsetLeftAndRight to control the position of the dragged Views. We want to use the same property to control the view position when it is animated, instead of dragged.