Only set offset when initializing connection
This is getting stranger, I still don't quite understand why if we set the offset here as well as tracking the offset, when the activity is recreated from dark mode back to light mode, the button will not be shown.

The symptom was that the `onLayout` is called twice in this situation, while 3 times from light to dark. This is caused by the fact that on dark mode we set a foregound to the FrameLayout, the additonal layout pass might be caused by it.

However, I still don't know why this will cause the button to not be rendered. The fix is likely not the right one, and but nevertheless it fixes the issue.