Switch back to setClickable for helper text (#104)
In a previous PR, we switched to setting the OnClickListener on both the helper text and the main button to avoid the view being able to be clicked when it should not be. The problem that we are having since then is that toggling helper text's clickable attribute became unnecessarily complex. When the helper text is clickable, it always goes to the about page.

Switching back to using setClickable to control whether the helper text is clickable, and only set the on click listener once.

https://ifttt-dev.atlassian.net/browse/MOB-838