Stop loading UI if suggested email is invalid
Instead of crashing the host app, we take a more graceful approach to deal with invalid emails on runtime: we will not render the UI if the email is invalid, but users can still stay in the app, but won't be able to use the ConnectButton.

An exception is that if the user has the app install, we bypass the email validation, because we will redirect users to the app instead of presenting email field.