Connect Button resume state fixes (#140)
* Make ConnectResult nullable

We actually need it to be nullable to represent cases where onNewIntent is called from non-redirect. In this case, we should not try to force the ConnectButton to be in either initial state or the email state.

* Reset knob background when running complete animation

To account for an issue caused by redirecting to Gmail: the app is temporarily moved back to foreground, and onResume is going to be called. With the ActivityLifecycleCallbacks registered, the onActivityOnResume is going to be called, and we are going to try to restore the button state, which is incorrect.

Adding an extra reset knob background step to account for this.