Start full flow if not authorized
This is to make sure the SDK is doing the right thing when the ConnectionApiClient is not authorized, i.e. there is no user token set in the header. In this case, we should not try to call the enable API directly but should go through the full connect flow.