Reset ConnectionApiClient's authorization when logout (#143)
* Reset ConnectionApiClient's authorization when logout

This is to demonstrate how we control the authorization for the API calls within ConnectButton, as well as fixing an issue in Grocery Epxress that the auth header is not cleared when users log out.

* Add notes in README

* Start full flow if not authorized

This is to make sure the SDK is doing the right thing when the ConnectionApiClient is not authorized, i.e. there is no user token set in the header. In this case, we should not try to call the enable API directly but should go through the full connect flow.