Connect Button re-enable flow fixes (#144)
* Redirect to web to resolve enable issues

Sometimes the SDK cached information might not be updated, and the re-enable flow might run into issues. We should redirect to web to let users go through the full flow to resolve any potential user token or service issues.

* Handle null ConnectResult on ConnectButton

We are handling tnull ConnectResult in BaseConnectButton, but forgot to do it in ConnectButton as well.