Use CredentialProvider in interceptor
Since the provider is tightly coupled with the API call authorization, I think it makes more sense to use it directly in the interceptor, to try to fetch the token at the moment when we are going to make any API call.

The benefit of doing this is that we remove the need for a separate AsyncTask to perform the getUserToken operation, as well as managing its state along with the lifecycle events. The caller of the ConnectionApi no longer need to worry about when it needs to use the AsyncTask to get user token, and we removes some references of the provider in some of our internal classes.

The downside of doing this is that now the CredentialProvider.getUserToken is going to be called potentially for every single API call. If the implementation of the method is not optimized, we are going to bring down the performance of the API client quite a bit.