Add a default button state to ConnectButton (#171)
I noticed that the behavior of the ButtonStateChangeListener is not consistent whene the button is set up, i.e it doesn't get invoked when the button is in its initial state, but will be invoked when the connection's status changes to enabled or disabled. This is because the default value for the button state is `Initial`.

This change adds a "unknown" state as the default value for a ConnectButton.

We should document this in the changelog.