Listen to reboot broadcast and refresh connection
Looks like Awareness API doesn't handle geofences re-registration after a device reboot.

To make sure host apps don't have to handle this, we can reuse the ConnectionRefresher (renamed to ConnectionRefreshWorker) that's scheduled to retrieve connection id data, and then enqueue a one time job immediately, to try to re-register all active geofences.

If the ConnectionRefreshWorker is not scheduled, the BroadcastReceiver will be no-op.