Update OkHttp version on all libraries
OkHttp v3.14.2 had a bug that will crash when building the Client if the device is running Android 11.

And because we are bumping OkHttp to 4+, we no longer support Android SDK lower than 21.

Also, along with the update, adding ACCESS_BACKGROUND_LOCATION requirement to ConnectLocation and mention it in README.