Move anonymous ID implementation to connect-api module
We are using anonymous ID in both connect-api and connect-button, both of which are for analytics purposes, and we also want them to be consistent. Moving the helper class from connect-button to connect-api to make sure we are using the same one for the header, the redirect URL and all of the analytics events.